+++ user_management/ui/views.py
@@ -1,8 +1,6 @@
 from django.conf import settings
 from django.contrib import messages
-from django.contrib.auth.views import redirect_to_login
 from django.core.exceptions import PermissionDenied
-from django.shortcuts import resolve_url
 from django.utils.translation import ugettext_lazy as _
 from django.views import generic
 
@@ -34,7 +32,6 @@ class VerifyUserEmailView(VerifyAccountViewMixin, generic.RedirectView):
         return super(VerifyUserEmailView, self).dispatch(request, *args, **kwargs)
 
     def get(self, request, *args, **kwargs):
-        next = resolve_url(settings.LOGIN_REDIRECT_URL)
         self.activate_user()
         messages.success(request, self.success_message)
-        return redirect_to_login(next, login_url=self.url)
+        return super(VerifyUserEmailView, self).get(request, *args, **kwargs)